Only have one task touch the ADC
Having more than one task touching the ADC creates a race condition and
it leads to missed / bad readings.

This change creates an array of axii and registers each axis with it
when the code starts up. The second reader task was removed in favor of
one task doing everything.

Also, I did some math and learned that a bit shift is just as accurate
as all of the floating point math I was doing, so I changed to that.
It's faster, but it doesn't really matter. 😅