Free up GPIOs by moving the buttons to an external mux
This works SO MUCH BETTER. It frees up a lot of GPIOs for me to do my
bidding with! I can even set up a second mux without needing another pin
(or use a 16 pin version). This requires an analog mux, which is more
expensive, but that's okay. I have a bunch on hand.

The buttons need a 5.1k pull-up resistor. They're driven to ground to
indicate that they've been pressed.